requirements.txt
# Gmail API
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Gemini AI
google-generativeai>=0.3.1

# Excel Operations
openpyxl>=3.1.2

# Git Operations
gitpython>=3.1.40

# CLI Interface
rich>=13.7.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.2
pyyaml>=6.0.1

# Retry Logic
tenacity>=8.2.3

# Development Dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1
